warning cleanup: misc fixes
- Arduino_LED_Matrix: fix packed attribute placement, add missing dependency
- Camera: fix pixel format check (formats are uint32_t, not pointers)
- Ethernet: fix float division literals
- SocketWrapper: hide deprecated warnings, fix signed/unsigned compare
- Wire: hide pedantic 'invalid offsetof' warning

+++ libraries/SocketWrapper/ZephyrUDP.h
@@ -269,7 +269,7 @@ private:
 	IPAddress _send_to_ip;
 	uint16_t _send_to_port;
 	std::vector<uint8_t> _tx_data;
-	int _rx_pkt_list_size = 10;
+	size_t _rx_pkt_list_size = 10;
 
 	/* UDP RECEPTION */
 	class UdpRxPacket {
